requirements.txt
pandas
matplotlib
torch
requests
pillow
ultralytics
opencv-python
scikit-image
numpy
wandb
streamlit